1/ PBFC2DSD15를 생성하는 코드 작성 (CoarseR = 0.1f, FineR = 0.025) Containter, breaking dam 생성하는 코드 수정함
- OpenMP사용시 velocityC나 constraintC를 계산할 때 문제가 발생하는 것으로 보임 > 디버그 필요함
 (OMP를 끈 상태로 training data 생성함)

2. visualization을 위한 코드 작성

+++ PIC2D/PIC2D.h
@@ -62,7 +62,7 @@ private:
 
 inline float linear_kernel(const Vector2f& d, const float& h)
 {
-	return std::max((1.0 - fabs(d(0) / h)) * (1.0 - fabs(d(1) / h)) * (1.0 - fabs(d(2) / h)), 0.0);
+	return float(std::max((1.0 - fabs(d(0) / h)) * (1.0 - fabs(d(1) / h)) * (1.0 - fabs(d(2) / h)), 0.0));
 }
 
 template<class T>
